refactor wallet restore tests
